requirements.txt
noise==1.2.2
opensimplex==0.3
Pillow==7.2.0
